the act_comm.cpp has been separated.
