+++ src/engine/ui/cmd/do_spec_comm.h
@@ -1,0 +1,17 @@
+/**
+\file do_spec_comm.h - a part of the Bylins engine.
+\authors Created by Sventovit.
+\date 24.09.2024.
+\brief Brief description.
+\detail Detail description.
+*/
+
+#ifndef BYLINS_SRC_ENGINE_UI_CMD_DO_SPEC_COMM_H_
+#define BYLINS_SRC_ENGINE_UI_CMD_DO_SPEC_COMM_H_
+
+class CharData;
+void do_spec_comm(CharData *ch, char *argument, int/* cmd*/, int subcmd);
+
+#endif //BYLINS_SRC_ENGINE_UI_CMD_DO_SPEC_COMM_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
